restructure directory and added check for release tag in python script
