+++ source/template.j2
@@ -56,7 +56,7 @@ We have made the following maintenance-related minor changes:
 {% endfor %}
 {% else %}
 {% if all_changes %}
-# Release [{{ change['release_version'] }}](https//github.com/nginx/agent/releases/tag/{{ change['release_version'] }})
+# Release [{{ change['release_version'] }}](https://github.com/nginx/agent/releases/tag/{{ change['release_version'] }})
 {% if change['changes']['🌟 Highlights'] %}
 ### 🌟 Highlights
 {% for highlight in change['changes']['🌟 Highlights'] %}
